# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateTadqiqotchi:
    properties:
      FanTarmogi:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      oqish_joyi:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
  models.UpdateTadqiqotchi:
    properties:
      FanTarmogi:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      oqish_joyi:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /tadqiqotchi:
    get:
      consumes:
      - application/json
      description: Get List Tadqiqotchi
      operationId: get_list_tadqiqotchi
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Tadqiqotchi
      tags:
      - Tadqiqotchi
    post:
      consumes:
      - application/json
      description: Create Tadqiqotchi
      operationId: create_tadqiqotchi
      parameters:
      - description: CreateTadqiqotchiRequest
        in: body
        name: tadqiqotchi
        required: true
        schema:
          $ref: '#/definitions/models.CreateTadqiqotchi'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Tadqiqotchi
      tags:
      - Tadqiqotchi
  /tadqiqotchi/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Tadqiqotchi
      operationId: delete_tadqiqotchi
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Tadqiqotchi
      tags:
      - Tadqiqotchi
    get:
      consumes:
      - application/json
      description: Get By ID Tadqiqotchi
      operationId: get_by_id_tadqiqotchi
      parameters:
      - description: id
        in: path
        name: id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get By ID Tadqiqotchi
      tags:
      - Tadqiqotchi
    put:
      consumes:
      - application/json
      description: Update Tadqiqotchi
      operationId: update_tadqiqotchi
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateTadqiqotchiRequest
        in: body
        name: tadqiqotchi
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTadqiqotchi'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Tadqiqotchi
      tags:
      - Tadqiqotchi
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"